Fix real time synchronization bug
git-svn-id: svn://svn.code.sf.net/p/powerdevs/code/trunk@683 005b60d0-fea2-4bd5-a63e-298cc19d6454

+++ engine/root_simulator.cpp
@@ -30,7 +30,7 @@
 Time realTi;
 bool overRun;
 int exitStatus=0;
-long int realTiSimulation;
+double realTiSimulation;
 
 void RootSimulator::init(){
 	t=ti;
